pyproject.toml
[project]
name = "modelo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "openai>=1.51.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.39.0",
]
